Introduce FileOperations and FileBasedProcessor
Add GitHubFileOperations and LocalFileOperations implementations and
refactor GitHubStore and LocalStore to reuse the shared file-based
processor. Add StoreError::InvalidPath and ParseError and export the
internal file operations modules.
